requirements.txt
# env config
python-decouple==3.0
django-cache-url==1.0.0
dj-database-url==0.3.0

# cache with redis
django-redis==4.3.0
hiredis==0.2.0

# sentry
sentry[postgres]==8.0.6

# ldap user store
django-auth-ldap==1.2.7

